Move and cleanup tests
